古いポリシーオブジェクトを `Deprecated` モジュール配下に置く
