+++ app/policies/deprecated/db/character_rows_form_policy.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-module Db
-  class OrganizationRowsFormPolicy < ApplicationPolicy
+module Deprecated::Db
+  class CharacterRowsFormPolicy < ApplicationPolicy
     def create?
       user.present? && user.committer?
     end
