+++ app/policies/deprecated/db/person_rows_form_policy.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-module Db
-  class CharacterRowsFormPolicy < ApplicationPolicy
+module Deprecated::Db
+  class PersonRowsFormPolicy < ApplicationPolicy
     def create?
       user.present? && user.committer?
     end
