+++ app/policies/deprecated/db/slot_rows_form_policy.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-module Db
-  class CastRowsFormPolicy < ApplicationPolicy
+module Deprecated::Db
+  class SlotRowsFormPolicy < ApplicationPolicy
     def create?
       user.present? && user.committer?
     end
